Create an event handler that will sort the food list High-Low Water Usage
