--- client/utils/sort-temp.js
@@ -1,7 +1,0 @@
-export default function sortHighLowWater (arr) {
-  return arr.sort((a, b) => {
-    if (a.waterUsage < b.waterUsage) return -1
-    else if (a.waterUsage > b.waterUsage) return 1
-    else return 0
-  })
-}
