+++ client/utils/sort.js
@@ -5,3 +5,11 @@ export function sortAZ (arr) {
     else return 0
   })
 }
+
+export function sortHighLowWater (arr) {
+  return arr.sort((a, b) => {
+    if (a.waterUsage < b.waterUsage) return -1
+    else if (a.waterUsage > b.waterUsage) return 1
+    else return 0
+  })
+}
