+++ tests/client/utils/sort.test.js
@@ -1,4 +1,4 @@
-import sortAZ from '../../../client/utils/sort'
+import { sortAZ, sortHighLowWater } from '../../../client/utils/sort'
 
 test('sortAZ sorts an array of objects alphabetically', () => {
   const foods = [
@@ -16,6 +16,26 @@ test('sortAZ sorts an array of objects alphabetically', () => {
   ]
 
   const actual = sortAZ(foods)
+
+  expect(actual).toStrictEqual(expected)
+})
+
+test('sortHighLowWater sorts an array of objects from high to low water usage', () => {
+  const foods = [
+    { id: 1, name: 'orange', waterUsage: 3000 },
+    { id: 2, name: 'apple', waterUsage: 2400 },
+    { id: 3, name: 'banana', waterUsage: 5530 },
+    { id: 4, name: 'carrot', waterUsage: 30 }
+  ]
+
+  const expected = [
+    { id: 4, name: 'carrot', waterUsage: 30 },
+    { id: 2, name: 'apple', waterUsage: 2400 },
+    { id: 1, name: 'orange', waterUsage: 3000 },
+    { id: 3, name: 'banana', waterUsage: 5530 }
+  ]
+
+  const actual = sortHighLowWater(foods)
 
   expect(actual).toStrictEqual(expected)
 })
